refactor(api): unify session endpoints and clean up API surface (#558)
Merge the two separate session systems (transport-specific read-only
endpoints in routes.ts and chat CRUD+SSE in chat.ts) into a single
unified /api/sessions endpoint set backed by sessionStoreManager.

Key changes:
- Create src/api/sessions.ts with unified session endpoints that read
  from sessionStoreManager.all() — any session is queryable regardless
  of origin (Discord, chat, etc.)
- Split routes.ts into cron.ts, config.ts, logs.ts, usage.ts
- Remove transportSessionStores indirection (redundant since plugins
  already register stores via sessionStoreManager)
- Remove registerSessionStores from plugin API
- Normalize all list responses to { items: [] } format
- Fix subagents param :taskId → :id
- Update TUI consumer to use unified /api/sessions paths
- Delete chat.ts (merged into sessions.ts)

Closes #558
