+++ src/plugins/manager.ts
@@ -15,7 +15,6 @@ import type {
   PluginConfigEntry,
   TransportFactory,
 } from "./types.js";
-import type { SessionStore } from "../core/types.js";
 
 const log = createLogger("plugins");
 
@@ -31,7 +30,6 @@ export class PluginManager {
   private uiRegistry = new UIRegistry();
   private transportFactories = new Map<string, TransportFactory>();
   private toolPluginRegistry = new ToolPluginRegistry();
-  private transportSessionStores = new Map<string, Map<string, SessionStore>>();
 
   async discoverAndLoad(
     searchDirs: string[],
@@ -78,7 +76,6 @@ export class PluginManager {
       uiRegistry: this.uiRegistry,
       transportFactories: this.transportFactories,
       toolPluginRegistry: this.toolPluginRegistry,
-      transportSessionStores: this.transportSessionStores,
       pluginConfig,
     });
 
@@ -102,10 +99,6 @@ export class PluginManager {
 
   getToolPluginRegistry(): ToolPluginRegistry {
     return this.toolPluginRegistry;
-  }
-
-  getTransportSessionStores(): Map<string, Map<string, SessionStore>> {
-    return this.transportSessionStores;
   }
 
   getLoadedPlugins(): PluginManifest[] {
